--- a/src/DDUS.jl
+++ b/src/DDUS.jl
@@ -4,7 +4,8 @@
 
 module DDUS
 
-using JuMPeR  #VG Talk to Iain about removing this
+# The JuMPeR oracle interface
+import JuMPeR: AbstractOracle, registerConstraint, setup, generateCut, generateReform
 
 include("helpers.jl")
 include("FBOracle.jl")
@@ -13,5 +14,8 @@ include("UCSOracle.jl")
 include("LCXOracle.jl")
 include("UMOracle.jl")
 #include("UDYOracle.jl")  #Needs SDP support in JuMPeR ...  add later
+
+# Implements the JuMPeR oracle interface for all oracles
+include("common.jl")
 
 end # module