--- a/src/FBOracle.jl
+++ b/src/FBOracle.jl
@@ -74,11 +74,6 @@ function suppFcnFB(xs, mfs, mbs, sigfs, sigbs, log_eps, cut_sense=:Max)
 end
 
 
-
-# JuMPeR alerting us that we're handling this contraint
-registerConstraint(w::FBOracle, rm::Model, ind::Int, prefs) = 
-	! get(prefs, :prefer_cuts, true) && error("Only cutting plane supported")
-
 function setup(w::FBOracle, rm::Model, prefs)
     # Extract preferences we care about
     w.debug_printcut = get(prefs, :debug_printcut, false)
@@ -92,40 +87,6 @@ function setup(w::FBOracle, rm::Model, prefs)
     @assert (length(rd.uncertaintyset) == 0) "Auxiliary constraints on uncertainties not yet supported"
 end
 
-function generateCut(w::FBOracle, m::Model, rm::Model, inds::Vector{Int}, active=false)
-    new_cons = {}
-    rd = JuMPeR.getRobust(rm)
 
-    for ind in inds
-        # Update the cutting plane problem's objective
-        con = JuMPeR.get_uncertain_constraint(rm, ind)
-        cut_sense, xs, lhs_const = JuMPeR.build_cut_objective(rm, con, m.colVal)
-        zstar, ustar = suppFcn(xs, w, cut_sense) 
-        lhs_of_cut = zstar + lhs_const
-
-        # SUBJECT TO CHANGE: active cut detection
-        if active
-            push!(rd.activecuts[ind], 
-                JuMPeR.cut_to_scen(ustar, 
-                    JuMPeR.check_cut_status(con, lhs_of_cut, w.cut_tol) == :Active))
-            continue
-        end
-
-        # Check violation
-        if JuMPeR.check_cut_status(con, lhs_of_cut, w.cut_tol) != :Violate
-            w.debug_printcut && JuMPeR.debug_printcut(rm ,m,w,lhs_of_cut,con,nothing)
-            continue  # No violation, no new cut
-        end
-        
-        # Create and add the new constraint
-        new_con = JuMPeR.build_certain_constraint(m, con, ustar)
-        w.debug_printcut && JuMPeR.debug_printcut(rm, m, w, lhs_of_cut, con, new_con)
-        push!(new_cons, new_con)
-    end
-    return new_cons
-end
-
-#Shouldn't be called
-generateReform(w::FBOracle, m::Model, rm::Model, inds::Vector{Int}) = 0