--- a/src/UMOracle.jl
+++ b/src/UMOracle.jl
@@ -1,11 +1,9 @@
 #####
 # UM
 #####
-import JuMPeR: registerConstraint, setup, generateCut, generateReform
 import JuMP.UnsetSolver
 
 export UMOracle
-export suppFcn
 
 #returns zstar, ustar
 function suppFcnUM(xs, lquants, uquants, cut_sense=:Max)
@@ -55,10 +53,6 @@ function UMOracle(data, lbounds, ubounds, eps_, alpha; cut_tol=1e-6, debug_print
 	end
 end
 
-# JuMPeR alerting us that we're handling this contraint
-registerConstraint(w::UMOracle, rm::Model, ind::Int, prefs) = 
-    ! get(prefs, :prefer_cuts, true) && error("Only cutting plane supported")
-
 function setup(w::UMOracle, rm::Model, prefs)
     # Extract preferences we care about
     w.debug_printcut = get(prefs, :debug_printcut, false)
@@ -68,44 +62,3 @@ function setup(w::UMOracle, rm::Model, prefs)
     @assert (rd.numUncs == length(w.uquants)) "Num Uncertainties $(rd.numUncs) doesn't match columns in uquants $(w.uquants)"
     @assert (length(rd.uncertaintyset) == 0) "Auxiliary constraints on uncertainties not yet supported"
 end
-
-#Refactor all this logic into a single place
-function generateCut(w::UMOracle, m::Model, rm::Model, inds::Vector{Int}, active=false)
-    new_cons = {}
-    rd = JuMPeR.getRobust(rm)
-    for ind in inds
-        con = JuMPeR.get_uncertain_constraint(rm, ind)
-        cut_sense, xs, lhs_const = JuMPeR.build_cut_objective(rm, con, m.colVal)
-        zstar, ustar = suppFcn(xs, w, cut_sense)
-        lhs_of_cut = zstar + lhs_const
-
-        # SUBJECT TO CHANGE: active cut detection
-        if active
-            push!(rd.activecuts[ind], 
-                JuMPeR.cut_to_scen(ustar, 
-                    JuMPeR.check_cut_status(con, lhs_of_cut, w.cut_tol) == :Active))
-            continue
-        end
-
-        # Check violation
-        if JuMPeR.check_cut_status(con, lhs_of_cut, w.cut_tol) != :Violate
-            w.debug_printcut && JuMPeR.debug_printcut(rm ,m,w,lhs_of_cut,con,nothing)
-            continue  # No violation, no new cut
-        end
-        
-        # Create and add the new constraint
-        new_con = JuMPeR.build_certain_constraint(m, con, ustar)
-        w.debug_printcut && JuMPeR.debug_printcut(rm, m, w, lhs_of_cut, con, new_con)
-        push!(new_cons, new_con)
-    end
-    return new_cons
-end
-
-#Shouldn't be called
-generateReform(w::UMOracle, m::Model, rm::Model, inds::Vector{Int}) = 0
-
-
-
-
-
-