--- a/src/common.jl
+++ b/src/common.jl
@@ -3,11 +3,11 @@ Generalized versions of the JuMPeR interface
 Delegates to correct cutting plane algorithm via suppFcn
 =#
 
-typealias DDUSOracles Union(UMOracle, UIOracle)
+typealias DDUSOracles Union(UMOracle, UIOracle, FBOracle, LCXOracle, UCSOracle)
 
 # JuMPeR alerting us that we're handling this contraint
-registerConstraint(w::DDUSOracles, rm::Model, ind::Int, prefs) = 
-    ! get(prefs, :prefer_cuts, true) && error("Only cutting plane supported")
+registerConstraint(w::DDUSOracles, rm::Model, ind::Int, prefs) =
+    !get(prefs, :prefer_cuts, true) && error("Only cutting plane supported")
 
 # JuMPeR wants us to generate a constraint for every uncertain constraint in inds
 function generateCut(w::DDUSOracles, m::Model, rm::Model, inds::Vector{Int}, active=false)