--- a/src/common.jl
+++ b/src/common.jl
@@ -2,7 +2,6 @@
 Generalized versions of the JuMPeR interface
 Delegates to correct cutting plane algorithm via suppFcn
 =#
-
 typealias DDUSOracles Union(UMOracle, UIOracle, FBOracle, LCXOracle, UCSOracle)
 
 # JuMPeR alerting us that we're handling this contraint