Goal: delete as much code as possible to make it easier to maintain and update.