@IainNZ 2015-01-23
Perf improvement for FBOracle:
```
before:
elapsed time: 0.652411114 seconds (416000000 bytes allocated, 45.20% gc time)
elapsed time: 0.413018565 seconds (288000000 bytes allocated, 39.18% gc time)
after:
elapsed time: 0.347826034 seconds (240000000 bytes allocated, 38.46% gc time)
elapsed time: 0.264211346 seconds (176000000 bytes allocated, 39.33% gc time)
```
Could be a bit faster by only storing `sigfs.^2` and `sigbs.^2`, because we never use the un-squared versions. Maybe a better test would be if the cut was longer than dim 2 but I don't know what the args mean.