@IainNZ 2015-01-23
**BIG** Perf improvement for UMOracle

```
before:
elapsed time: 4.276407198 seconds (1512001088 bytes allocated, 18.91% gc time)
elapsed time: 4.330083944 seconds (1512000000 bytes allocated, 18.51% gc time)
after:
elapsed time: 0.186735935 seconds (176000000 bytes allocated, 48.54% gc time)
elapsed time: 0.152339507 seconds (176000000 bytes allocated, 40.06% gc time)
```

Removed the creation of the three temporary arrays in the `sum(max` block, and just generally tried to shorten the code while I was in there.