@IainNZ 2015-01-23
UIOracle

```
before:
elapsed time: 5.983563354 seconds (5642481088 bytes allocated, 53.91% gc time)
elapsed time: 5.885752918 seconds (5641840000 bytes allocated, 54.09% gc time)
after:
elapsed time: 2.357058779 seconds (535761088 bytes allocated, 13.71% gc time)
elapsed time: 2.363098495 seconds (535760000 bytes allocated, 13.79% gc time)
```

But waiting on fix for final verdict.