@IainNZ 2015-01-25
Will merge in once Travis passes